ipmi: ensure forward progress on ipmi_queue_msg_sync()
BT responses are handled using a timer doing the polling. To hope to
get an answer to an IPMI synchronous message, the timer needs to run.

We can't just check all timers though as there may be a timer that
wants a lock that's held by a code path calling ipmi_queue_msg_sync(),
and if we did enforce that as a requirement, it's a pretty subtle
API that is asking to be broken.

So, if we just run a poll function to crank anything that the IPMI
backend needs, then we should be fine.

This issue shows up very quickly under QEMU when loading the first
flash resource with the IPMI HIOMAP backend.

Reported-by: Cédric Le Goater <clg@kaod.org>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>
Reviewed-by: Andrew Jeffery <andrew@aj.id.au>
Reviewed-by: Cédric Le Goater <clg@kaod.org>
Reviewed-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
(cherry picked from commit f01cd777adb16cbab93215d26159aa1c4606112c)
Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>