core/pci: Train link of PHB slots when hotplugging
The link of PHB slots must be trained after powering on. This can be
done by calling the fundamental reset callback of the slot.

We could force a reset for all the slots and have a common path in
set_power_state(). But this patch only resets the PHB slot. Some slot
implementations do a power cycle during fundamental reset, so calling
a reset after powering on would repeat that operation.

Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>